+++ playground/server/routes/bull-board.ts
@@ -1,4 +1,0 @@
-import { defineEventHandler } from 'h3'
-import { redirectToBullboard } from '../handlers/bull-board'
-
-export default defineEventHandler(redirectToBullboard)